fix-feat: upload image product and dosen

+++ src/pages/api/product/[[...product]].ts
@@ -1,5 +1,6 @@
 import {
   addData,
+  addDataId,
   deleteData,
   retrieveData,
   updateData,
@@ -19,67 +20,81 @@ export default async function handler(
     jwt.verify(
       token,
       process.env.NEXTAUTH_SECRET || "",
-      async (err: any, decoded: any) => {
-        console.log(decoded);
+      async (error: any, decoded: any) => {
         if (decoded && decoded.role === "admin") {
           let data = req.body;
-          data.createdAt = new Date().toDateString();
-          data.updatedAt = new Date().toDateString();
-          console.log(data);
-          addData("news", data, (status: boolean, result: any) => {
+          data.createdAt = new Date().getTime();
+          data.updatedAt = new Date().getTime();
+
+          await addData("news", data, (status: boolean, id: string) => {
             if (status) {
               res.status(200).json({
+                status: true,
                 message: "success",
                 statusCode: 200,
-                status: true,
-                data: { id: result.id },
+                data: { id: id },
               });
             } else {
-              res.status(400).json({
+              res.status(402).json({
                 status: false,
                 message: "failed",
-                statusCode: 400,
+                statusCode: 402,
                 data: {},
               });
             }
           });
-        } else {
-          res.status(403).json({ message: "failed", statusCode: 403 });
         }
       }
     );
+    // res.status(200).json({ message: "POST handler", statusCode: 200 });
   } else if (req.method === "PUT") {
     const { product }: any = req.query;
-    console.log(product);
     const { data } = req.body;
     console.log(product);
+
     const token = req.headers.authorization?.split(" ")[1] || "";
     jwt.verify(
       token,
       process.env.NEXTAUTH_SECRET || "",
-      async (err: any, decoded: any) => {
+      async (error: any, decoded: any) => {
         if (decoded && decoded.role === "admin") {
           await updateData("news", product[0], data, (status: boolean) => {
+            if (status) {
+              res
+                .status(200)
+                .json({ status: true, message: "success", statusCode: 200 });
+            } else {
+              res
+                .status(400)
+                .json({ status: false, message: "failed", statusCode: 400 });
+            }
+          });
+        } else {
+          res
+            .status(400)
+            .json({ status: false, message: "access denied", statusCode: 400 });
+        }
+      }
+    );
+    res.status(200).json({ message: "PUT handler", statusCode: 200 });
+  } else if (req.method === "DELETE") {
+    const { product }: any = req.query;
+    const token = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (error: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
+          await deleteData("news", product[0], (status: boolean) => {
             if (status) {
               res.status(200).json({ message: "success", statusCode: 200 });
             } else {
               res.status(400).json({ message: "failed", statusCode: 400 });
             }
           });
-        } else {
-          res.status(403).json({ message: "failed", statusCode: 403 });
         }
       }
     );
-  } else if (req.method === "DELETE") {
-    const { product }: any = req.query;
-    console.log({ product });
-    deleteData("news", product[0], (status: boolean) => {
-      if (status) {
-        res.status(209).json({ message: "success", statusCode: 209 });
-      } else {
-        res.status(409).json({ message: "failed", statusCode: 409 });
-      }
-    });
+    res.status(200).json({ message: "DELETE handler", statusCode: 200 });
   }
 }
